pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-composable-installer"
version = "0.1.0"
description = "AumOS Enterprise one-installation, selective-activation platform installer"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "AumOS Enterprise" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Installation/Setup",
]
dependencies = [
    "aumos-common>=0.1.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "structlog>=24.1.0",
]

[project.scripts]
aumos = "aumos_composable_installer.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aumos_composable_installer"]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "PT", "RUF"]
ignore = ["ANN101", "ANN102", "COM812"]

[tool.ruff.lint.isort]
known-first-party = ["aumos_composable_installer"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=aumos_composable_installer --cov-report=term-missing"
